refactor(agent-shared): move columnar encoder into shared package
Move parseFields, parseLimit, projectRows, shapePaginatedRows, and
shapeToolResult from packages/cli into @rozenite/agent-shared so MCP and
other agent-sdk consumers can reuse the same envelope, per issue #318's
maintainer comment. shapePaginatedRows/shapeToolResult now take a
frontend-neutral `nextAffordance` string instead of a CLI-specific
`nextCommand`, so the shared functions carry no shell-command assumptions.

paginateRows (cursor encode/decode) and formatAgentCommand/shellEscape
stay CLI-only: paginateRows is an offset pager scoped to CLI-owned
tools/domains listings, not the general shared pagination engine that
issue #320 already owns, and the shell-command builder is explicitly a
CLI-only rendering of the next-page affordance per the maintainer's
carve-out.

Adds a test script + vitest suite to agent-shared, which previously had
none.
