+++ packages/agent-shared/README.md
@@ -10,6 +10,7 @@
 
 - **Shared Tool Types**: Common `AgentTool` and JSON-schema-like input types
 - **Paginated Tool Contracts**: Typed row fields and a shared page envelope for built-in and third-party tools
+- **Row Shaping**: Frontend-neutral columnar/row-keyed encoding (`shapePaginatedRows`, `shapeToolResult`) shared by the CLI and any other Agent frontend
 - **Message Contracts**: Typed payloads for register, unregister, call, and result messages
 - **Single Protocol Constant**: Shared `AGENT_PLUGIN_ID` for the Agent transport
 - **Package Reuse**: Intended for bridge and runtime packages that implement Agent support
@@ -32,6 +33,13 @@ This package exports:
 - `PageEnvelope`
 - `PageResult`
 - `definePaginatedAgentToolContract`
+- `DEFAULT_PAGE_LIMIT`
+- `MAX_PAGE_LIMIT`
+- `parseFields`
+- `parseLimit`
+- `projectRows`
+- `shapePaginatedRows`
+- `shapeToolResult`
 - `JSONSchema7`
 - `DevToolsPluginMessage`
 - `RegisterToolPayload`
